fix: keep release smoke independent of signing secrets
