+++ forge.config.ts
@@ -5,6 +5,7 @@ import { flipFuses, FuseV1Options, FuseVersion } from "@electron/fuses";
 import { readFileSync, writeFileSync } from "node:fs";
 import path from "node:path";
 import { macOSBundleVersions } from "./scripts/macos-version.js";
+import { ignorePackageFile } from "./scripts/package-ignore.js";
 import { resolvePackageMode } from "./scripts/package-mode.js";
 import {
   distributionEntitlementsPath,
@@ -85,30 +86,7 @@ const config: ForgeConfig = {
       NSAppTransportSecurity: { NSAllowsArbitraryLoads: false },
     },
     // Forge's own packaged output is out/; compiled JS lives in build/.
-    ignore: (file) => {
-      if (!file || file === "/") return false;
-      const p = file.startsWith("/") ? file : `/${file}`;
-      if (p === "/package.json") return false;
-      if (p === "/build" || p === "/build/main" || p === "/build/shared")
-        return false;
-      if (p.startsWith("/build/main/") || p.startsWith("/build/shared/")) {
-        return (
-          p.endsWith(".map") || p.endsWith(".d.ts") || p.endsWith(".d.ts.map")
-        );
-      }
-      if (p === "/build/renderer") return false;
-      if (p.startsWith("/build/renderer/")) return p.endsWith(".d.ts");
-      if (p === "/build/preload" || p === "/build/preload/preload.cjs")
-        return false;
-      if (
-        p === "/build/native"
-        || p === "/build/native/keychain.node"
-        || p === "/build/native/gw-dat-decode"
-      ) {
-        return false;
-      }
-      return true;
-    },
+    ignore: ignorePackageFile,
   },
   rebuildConfig: {},
   makers: [
